sonar nits
